% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{attendance_df}
\alias{attendance_df}
\title{Hourly attendance data for Lubaszka Solec bakery}
\format{
A data frame with 14 rows and 7 columns: \code{hour}, \code{mon},
\code{tue}, \code{wed}, \code{thu}, \code{fri}, \code{sat}.
}
\source{
\url{https://maps.app.goo.gl/JhYvKgQEy5hVBCzB9}
}
\usage{
attendance_df
}
\description{
Relative customer attendance per hour from 6h to 20h
(1-hour intervals: 6h-7h, 7h-8h, ..., 19h-20h)
for each weekday, sourced from Google Maps Popular Times.
}
