% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{Bread-class}
\alias{Bread-class}
\alias{Bread}
\title{Bread: product subclass for bakery breads}
\description{
Inherits all slots and validity from \linkS4class{Product}.
Sets a morning-weighted demand profile and dawn production time.
}
