% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{chleb_fitness}
\alias{chleb_fitness}
\title{Chleb fitness product}
\format{
An S4 object of class \code{Bread}.
}
\usage{
chleb_fitness
}
\description{
Chleb fitness product
}
\keyword{datasets}
