% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{chleb_grycany}
\alias{chleb_grycany}
\title{Chleb gryczany product}
\format{
An S4 object of class \code{Bread}.
}
\usage{
chleb_grycany
}
\description{
Chleb gryczany product
}
\keyword{datasets}
