% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{chleb_wiejski}
\alias{chleb_wiejski}
\title{Chleb wiejski product}
\format{
An S4 object of class \code{Bread}.
}
\usage{
chleb_wiejski
}
\description{
Chleb wiejski product
}
\keyword{datasets}
