% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kanapka_reuben}
\alias{kanapka_reuben}
\title{Kanapka Reuben product}
\format{
An S4 object of class \code{Sandwich}.
}
\usage{
kanapka_reuben
}
\description{
Kanapka Reuben product
}
\keyword{datasets}
