% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kanapka_weganska}
\alias{kanapka_weganska}
\title{Kanapka weganska product}
\format{
An S4 object of class \code{Sandwich}.
}
\usage{
kanapka_weganska
}
\description{
Kanapka weganska product
}
\keyword{datasets}
