% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lubaszka_solec}
\alias{lubaszka_solec}
\title{Lubaszka Solec bakery object}
\format{
An S4 object of class \code{Bakery}.
}
\usage{
lubaszka_solec
}
\description{
Main S4 Bakery object representing the Lubaszka bakery (Solec location),
initialized with real hourly attendance data.
}
\keyword{datasets}
