% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{Pizza-class}
\alias{Pizza-class}
\alias{Pizza}
\title{Pizza: product subclass for bakery pizzas}
\description{
Inherits all slots and validity from \linkS4class{Product}.
Sets a lunch and early-afternoon demand profile with noon production time.
}
