% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pizza_margherita}
\alias{pizza_margherita}
\title{Pizza Margherita product}
\format{
An S4 object of class \code{Pizza}.
}
\usage{
pizza_margherita
}
\description{
Pizza Margherita product
}
\keyword{datasets}
