% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{product_catalog}
\alias{product_catalog}
\title{Product catalog}
\format{
A named list of S4 objects (Sandwich, Bread, Pizza).
}
\usage{
product_catalog
}
\description{
Named list of all S4 product objects, used for Shiny input selection.
}
\keyword{datasets}
