% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{Sandwich-class}
\alias{Sandwich-class}
\alias{Sandwich}
\title{Sandwich: product subclass for bakery sandwiches}
\description{
Inherits all slots and validity from \linkS4class{Product}.
Sets a lunch-peaked demand profile and mid-morning production time.
}
